Introduce a simple interface that allows moving LimitActiveUsersSettings to core. Tweak API action.
